Update thhe file structure
