+++ src/hooks/documents/useAddDocument.ts
@@ -1,0 +1,24 @@
+import axios from "axios";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+
+const addDocument = async (formData: FormData) => {
+  const response = await axios.post("/api/documents/upload", formData);
+
+  return response.data;
+};
+
+const useAddDocument = () => {
+  const queryClient = useQueryClient();
+
+  return useMutation({
+    mutationFn: addDocument,
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ["documents"] });
+    },
+    onError: (error) => {
+      console.error("Error adding document: ", error);
+    }
+  });
+};
+
+export default useAddDocument;
